Rename files and classes to more meaningfull names
