+++ qredmineclient_global.h
